Made a start with predictive analysis
I set up the pipeline for training predictive models on the historical KPI (X) Per-level (Y) data and I set up a KNN algorithm and a baseline which always selects the most frequent Y label. The KNN-model performs exceptionally well, especially for k=1, but this is probably just because days that are next to each other (e.g. 21-03-2021 and 22-03-2021) tend to have similair kpi and PER values, so the model is likely to just always predict that PERs will stay at the same level, which is quite useless.